% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{construct_series_code}
\alias{construct_series_code}
\title{Construct ECB series code from series key}
\usage{
construct_series_code(series_key)
}
\arguments{
\item{series_key}{Character. An ECB series key.}
}
\value{
Character. The constructed series code.
}
\description{
Converts an ECB series key (e.g., "ICP.M.U2.N.000000.4.ANR") to the
database series code format (e.g., "ECB--ICP--U2--N--000000--4--ANR--M").
}
