% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{ECB_import_structure}
\alias{ECB_import_structure}
\title{Import structure for a single ECB series (dimensions, levels, series etc)}
\usage{
ECB_import_structure(
  series_key,
  con,
  source_id = 8,
  schema = "platform",
  keep_vintage = FALSE
)
}
\arguments{
\item{series_key}{Character. An ECB series key (e.g., "ICP.M.U2.N.000000.4.ANR").}

\item{con}{Database connection object}

\item{source_id}{Integer source ID, default 8 for ECB}

\item{schema}{defautls to "platform"}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}
}
\description{
NB: we are not using categories for the ECB dataflows since the website doesn't
have a clean hierarchy.
}
