% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_key_dimensions}
\alias{extract_key_dimensions}
\title{Extract key dimensions from ECB series key}
\usage{
extract_key_dimensions(key, dims_result)
}
\arguments{
\item{key}{Character, ECB series key}

\item{dims_result}{Output from ecb::get_dimensions(key)}
}
\value{
Data frame with only key dimensions (dim, value columns)
}
\description{
Filters get_dimensions() output to only include actual key dimensions,
excluding attributes.
}
\keyword{internal}
