% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_ecb_data_for_insert}
\alias{prepare_ecb_data_for_insert}
\title{Prepare ECB data  for insertion}
\usage{
prepare_ecb_data_for_insert(series_key, con, schema = "platform")
}
\arguments{
\item{series_key}{Character. An ECB series key (e.g., "ICP.M.U2.N.000000.4.ANR").}

\item{con}{Database connection}

\item{schema}{the schema to use for the connection, default is "platform"}
}
\value{
A list containing:
 - data: The processed data frame. which has to have columns time and value,
 with the rest corresponding to the dimensions of the table/series
 - table_id: The table
 - interval_id: the interval id ("M", "A", or "Q")
 - dimension_ids: The non-time dimension IDs
 - dimension_names: The names of the dimensions
}
\description{
Processes raw ECB data into a format ready for database insertion - this
is a single series funciton as opposed to most other fetchR prep_data functions.
}
