% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.R
\name{ECB_import_data_points}
\alias{ECB_import_data_points}
\title{Insert data points from BS}
\usage{
ECB_import_data_points(series_key, con, schema = "platform")
}
\arguments{
\item{series_key}{Character, ecb dataset code}

\item{con}{Database connection}

\item{schema}{Schema name}
}
\value{
Insertion results (invisibly)
}
\description{
Function to prepare and insert BS data points. The function first prepares
the required vintages and inserts them, then prepares the data points
table and inserts it. The function returns the results invisibly.
}
\details{
This is a BS specific function, which should be followed by the generic
UMARimportR function to write the vintage hashes and clean up redundant
vintages.
}
